Creates function to display only untaken surveys
Creates `dashboard` function to create `mySurveys` and `takeableSurveys`
arrays. This function is invoked with `data` passed to it from the
original `getAllSurveysSuccess` function call.

The resulting arrays are returned and passed to handlebars to display.

This is not an ideal method for accomplishing this goal. The `takenIdList`
array should be something implemented server side and updated when a
submission is successfully created. This should be added in a future
update